apps/nemo-guard/requirements.txt
nemoguardrails==0.9.1
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
httpx==0.27.0
python-dotenv==1.0.1
# Required by NeMo for async LLM calls
openai==1.35.0
# Structured logging
python-json-logger==2.0.7
